Added tableRows, and better debugging.
Added support for the new complex table rows format. This is not
designed for humans to do anything with, and it is not helpful
for anything but displaying data in nested tables. It is also
unfortunately difficult to predict the structure of.

A debug flag is added to the service to support log messages that
can be left in the code, and lessen the chance of things slipping
through.