Fixed list making.
List making was rather broken due to the failure to carry over implicit
join constraints. The procedure for making the list query has been
refactored out from  those methods that use it and improved to ensure
that results are of the same size as a row count for a query.